// frontend/src/main/model/LocationAssociationName.ts
export enum LocationAssociationName {
  COVER = 'COVER',
  PROFILE = 'PROFILE',
  LOGO = 'LOGO',
  EXTERIOR = 'EXTERIOR',
  INTERIOR = 'INTERIOR',
  PRODUCT = 'PRODUCT',
  AT_WORK = 'AT_WORK',
  FOOD_AND_DRINK = 'FOOD_AND_DRINK',
  MENU = 'MENU',
  ROOMS = 'ROOMS',
  TEAMS = 'TEAMS',
  ADDITIONAL = 'ADDITIONAL',
  CATEGORY_UNSPECIFIED = 'CATEGORY_UNSPECIFIED',
}